+++ autokeras/image/image_supervised.py
@@ -1,5 +1,5 @@
 import os
-from abc import abstractmethod
+from abc import abstractmethod, ABC
 from functools import reduce
 
 import numpy as np
@@ -11,7 +11,7 @@ from autokeras.constant import Constant
 from autokeras.nn.loss_function import classification_loss, regression_loss
 from autokeras.nn.metric import Accuracy, MSE
 from autokeras.preprocessor import OneHotEncoder, ImageDataTransformer
-from autokeras.supervised import Supervised, PortableClass
+from autokeras.supervised import Supervised, PortableClass, DeepSupervised
 from autokeras.utils import has_file, pickle_from_file, pickle_to_file, temp_folder_generator, validate_xy, \
     read_csv_file, read_image, compute_image_resize_params, resize_image_data
 
@@ -63,7 +63,7 @@ def load_image_dataset(csv_file_path, images_path):
     return np.array(x), np.array(y)
 
 
-class ImageSupervised(Supervised):
+class ImageSupervised(DeepSupervised, ABC):
     """Abstract image supervised class.
 
     Attributes:
@@ -80,7 +80,7 @@ class ImageSupervised(Supervised):
         resize_height: resize image height.
         resize_width: resize image width.
     """
-    def __init__(self, verbose=False, path=None, resume=False, searcher_args=None, augment=None):
+    def __init__(self, augment=None, **kwargs):
         """Initialize the instance.
         The classifier will be loaded from the files in 'path' if parameter 'resume' is True.
         Otherwise it would create a new one.
@@ -93,44 +93,17 @@ class ImageSupervised(Supervised):
             augment: A boolean value indicating whether the data needs augmentation. If not define, then it
                 will use the value of Constant.DATA_AUGMENTATION which is True by default.
         """
-        super().__init__(verbose)
-
-        if searcher_args is None:
-            searcher_args = {}
-
-        if path is None:
-            path = temp_folder_generator()
-
         if augment is None:
             augment = Constant.DATA_AUGMENTATION
-
-        self.path = path
-        if has_file(os.path.join(self.path, 'classifier')) and resume:
-            classifier = pickle_from_file(os.path.join(self.path, 'classifier'))
-            self.__dict__ = classifier.__dict__
-            self.cnn = pickle_from_file(os.path.join(self.path, 'module'))
-        else:
-            self.y_encoder = None
-            self.data_transformer = None
-            self.verbose = verbose
-            self.augment = augment
-            self.cnn = CnnModule(self.loss, self.metric, searcher_args, path, verbose)
-
+        self.augment = augment
         self.resize_height = None
         self.resize_width = None
 
-    @property
-    @abstractmethod
-    def metric(self):
-        pass
-
-    @property
-    @abstractmethod
-    def loss(self):
-        pass
+        super().__init__(**kwargs)
 
     def fit(self, x, y, x_test=None, y_test=None, time_limit=None):
         x = np.array(x)
+        y = np.array(y).flatten()
 
         if len(x.shape) != 0 and len(x[0].shape) == 3:
             if self.verbose:
@@ -142,42 +115,11 @@ class ImageSupervised(Supervised):
             if self.verbose:
                 print("Preprocessing finished.")
 
-        y = np.array(y).flatten()
-        validate_xy(x, y)
-        y = self.transform_y(y)
-        if x_test is None or y_test is None:
-            # Divide training data into training and testing data.
-            validation_set_size = int(len(y) * Constant.VALIDATION_SET_SIZE)
-            validation_set_size = min(validation_set_size, 500)
-            validation_set_size = max(validation_set_size, 1)
-            x_train, x_test, y_train, y_test = train_test_split(x, y,
-                                                                test_size=validation_set_size,
-                                                                random_state=42)
-        else:
-            x_train = x
-            y_train = y
-        # Transform x_train
+        super().fit(x, y, x_test, y_test, time_limit)
+
+    def init_transformer(self, x):
         if self.data_transformer is None:
             self.data_transformer = ImageDataTransformer(x, augment=self.augment)
-
-        # Wrap the data into DataLoaders
-        train_data = self.data_transformer.transform_train(x_train, y_train)
-        test_data = self.data_transformer.transform_test(x_test, y_test)
-
-        # Save the classifier
-        pickle_to_file(self, os.path.join(self.path, 'classifier'))
-
-        if time_limit is None:
-            time_limit = 24 * 60 * 60
-
-        self.cnn.fit(self.get_n_output_node(), x_train.shape, train_data, test_data, time_limit)
-
-    @abstractmethod
-    def get_n_output_node(self):
-        pass
-
-    def transform_y(self, y_train):
-        return y_train
 
     def predict(self, x_test):
         """Return predict results for the testing data.
@@ -201,9 +143,6 @@ class ImageSupervised(Supervised):
         output = reduce(lambda x, y: np.concatenate((x, y)), outputs)
         return self.inverse_transform_y(output)
 
-    def inverse_transform_y(self, output):
-        return output
-
     def evaluate(self, x_test, y_test):
         """Return the accuracy score between predict value and `y_test`."""
         if len(x_test.shape) != 0 and len(x_test[0].shape) == 3:
@@ -222,13 +161,13 @@ class ImageSupervised(Supervised):
             trainer_args: A dictionary containing the parameters of the ModelTrainer constructor.
             retrain: A boolean of whether reinitialize the weights of the model.
         """
-        if trainer_args is None:
-            trainer_args = {'max_no_improvement_num': 30}
-
         if len(x_train.shape) != 0 and len(x_train[0].shape) == 3:
             x_train = resize_image_data(x_train, self.resize_height, self.resize_width)
             if x_test is not None:
                 x_test = resize_image_data(x_test, self.resize_height, self.resize_width)
+
+        if trainer_args is None:
+            trainer_args = {'max_no_improvement_num': 30}
 
         y_train = self.transform_y(y_train)
         y_test = self.transform_y(y_test)
