+++ autokeras/text/text_supervised.py
@@ -1,22 +1,21 @@
 import os
-import pickle
+from abc import ABC
 from functools import reduce
 
 import numpy as np
 import torch
 from sklearn.model_selection import train_test_split
 
-from autokeras.net_module import CnnModule
 from autokeras.constant import Constant
 from autokeras.nn.loss_function import classification_loss, regression_loss
 from autokeras.nn.metric import Accuracy, MSE
 from autokeras.preprocessor import OneHotEncoder, TextDataTransformer
-from autokeras.supervised import Supervised
+from autokeras.supervised import DeepSupervised
 from autokeras.text.text_preprocessor import text_preprocess
-from autokeras.utils import pickle_to_file, validate_xy, temp_folder_generator, has_file, pickle_from_file
+from autokeras.utils import pickle_to_file, validate_xy
 
 
-class TextClassifier(Supervised):
+class TextSupervised(DeepSupervised, ABC):
     """TextClassifier class.
 
     Attributes:
@@ -28,7 +27,8 @@ class TextClassifier(Supervised):
         verbose: A boolean value indicating the verbosity mode which determines whether the search process
                 will be printed to stdout.
     """
-    def __init__(self, verbose=False, path=None, resume=False, searcher_args=None):
+
+    def __init__(self, **kwargs):
         """Initialize the instance.
 
         The classifier will be loaded from the files in 'path' if parameter 'resume' is True.
@@ -40,24 +40,7 @@ class TextClassifier(Supervised):
                 Otherwise, the classifier will start a new search.
             searcher_args: A dictionary containing the parameters for the searcher's __init__ function.
         """
-        super().__init__(verbose)
-
-        if searcher_args is None:
-            searcher_args = {}
-
-        if path is None:
-            path = temp_folder_generator()
-
-        self.cnn = CnnModule(self.loss, self.metric, searcher_args, path, verbose)
-
-        self.path = path
-        if has_file(os.path.join(self.path, 'text_classifier')) and resume:
-            classifier = pickle_from_file(os.path.join(self.path, 'text_classifier'))
-            self.__dict__ = classifier.__dict__
-        else:
-            self.y_encoder = None
-            self.data_transformer = None
-            self.verbose = verbose
+        super().__init__(**kwargs)
 
     def fit(self, x, y, x_test=None, y_test=None, batch_size=None, time_limit=None):
         """Find the best neural architecture and train it.
@@ -77,37 +60,14 @@ class TextClassifier(Supervised):
         x = text_preprocess(x, path=self.path)
 
         x = np.array(x)
-        y = np.array(y)
-        validate_xy(x, y)
-        y = self.transform_y(y)
+        y = np.array(y).flatten()
 
-        if batch_size is None:
-            batch_size = Constant.MAX_BATCH_SIZE
-        # Divide training data into training and testing data.
-        if x_test is None or y_test is None:
-            x_train, x_test, y_train, y_test = train_test_split(x, y,
-                                                                test_size=min(Constant.VALIDATION_SET_SIZE,
-                                                                              int(len(y) * 0.2)),
-                                                                random_state=42)
-        else:
-            x_train = x
-            y_train = y
+        super().fit(x, y, x_test, y_test, time_limit)
 
+    def init_transformer(self, x):
         # Wrap the data into DataLoaders
         if self.data_transformer is None:
             self.data_transformer = TextDataTransformer()
-
-        train_data = self.data_transformer.transform_train(x_train, y_train, batch_size=batch_size)
-        test_data = self.data_transformer.transform_test(x_test, y_test)
-
-        # Save the classifier
-        pickle.dump(self, open(os.path.join(self.path, 'text_classifier'), 'wb'))
-        pickle_to_file(self, os.path.join(self.path, 'text_classifier'))
-
-        if time_limit is None:
-            time_limit = 24 * 60 * 60
-
-        self.cnn.fit(self.get_n_output_node(), x_train.shape, train_data, test_data, time_limit)
 
     def final_fit(self, x_train=None, y_train=None, x_test=None, y_test=None, trainer_args=None, retrain=False):
         """Final training after found the best architecture.
@@ -120,23 +80,18 @@ class TextClassifier(Supervised):
             trainer_args: A dictionary containing the parameters of the ModelTrainer constructor.
             retrain: A boolean of whether reinitialize the weights of the model.
         """
+        x_train = text_preprocess(x_train, path=self.path)
+        if x_test is not None:
+            x_test = text_preprocess(x_test, path=self.path)
+
         if trainer_args is None:
             trainer_args = {'max_no_improvement_num': 30}
-
-        if x_test is None:
-            x_train, x_test, y_train, y_test = train_test_split(x_train, y_train,
-                                                                test_size=min(Constant.VALIDATION_SET_SIZE,
-                                                                              int(len(y_train) * 0.2)),
-                                                                random_state=42)
-
-        x_train = text_preprocess(x_train, path=self.path)
-        x_test = text_preprocess(x_test, path=self.path)
 
         y_train = self.transform_y(y_train)
         y_test = self.transform_y(y_test)
 
-        train_data = self.data_transformer.transform_train(x_train, y_train, batch_size=Constant.MAX_BATCH_SIZE)
-        test_data = self.data_transformer.transform_test(x_test, y_test, batch_size=Constant.MAX_BATCH_SIZE)
+        train_data = self.data_transformer.transform_train(x_train, y_train)
+        test_data = self.data_transformer.transform_test(x_test, y_test)
 
         self.cnn.final_fit(train_data, test_data, trainer_args, retrain)
 
@@ -168,6 +123,8 @@ class TextClassifier(Supervised):
         y_predict = self.predict(x_test)
         return self.metric().evaluate(y_test, y_predict)
 
+
+class TextClassifier(TextSupervised):
     @property
     def metric(self):
         return Accuracy
@@ -187,9 +144,6 @@ class TextClassifier(Supervised):
     def inverse_transform_y(self, output):
         return self.y_encoder.inverse_transform(output)
 
-    def load_searcher(self):
-        return pickle_from_file(os.path.join(self.path, 'searcher'))
-
     def get_n_output_node(self):
         return self.y_encoder.n_classes
 
@@ -200,6 +154,7 @@ class TextRegressor(TextClassifier):
     It is used for text regression. It searches convolutional neural network architectures
     for the best configuration for the text dataset.
     """
+
     @property
     def loss(self):
         return regression_loss
